Move controller api key getter to utils
